requirements.txt
ultralytics==8.2.103
opencv-python==4.10.0.84
pyyaml==6.0.2
pydantic==2.8.2
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn==0.30.3
pandas==2.2.2
streamlit==1.37.1
plotly==5.23.0
fpdf==1.7.2
kaleido==0.2.1
